requirements.txt
streamlit
feedparser
pandas
requests
python-docx
textblob
altair
pytz
xlsxwriter
openpyxl
